Revert custom file drop implementation
