+++ src-ui/src/app/app.component.html
@@ -1,10 +1,16 @@
 <pngx-toasts></pngx-toasts>
 
-<pngx-file-drop>
-    <ng-container content>
-        <router-outlet></router-outlet>
-    </ng-container>
-</pngx-file-drop>
+<ngx-file-drop dropZoneClassName="main-dropzone" contentClassName="main-content" [disabled]="!dragDropEnabled"
+(onFileDrop)="dropped($event)" (onFileOver)="fileOver()" (onFileLeave)="fileLeave()">
+    <ng-template ngx-file-drop-content-tmp>
+        <div class="global-dropzone-overlay fade" [class.show]="fileIsOver" [class.hide]="hidden">
+            <h2 i18n>Drop files to begin upload</h2>
+        </div>
+        <div [class.inert]="fileIsOver">
+            <router-outlet></router-outlet>
+        </div>
+    </ng-template>
+</ngx-file-drop>
 
 <tour-step-template>
     <ng-template #tourStep let-step="step">
