--- src-ui/src/app/components/file-drop/file-drop.component.ts
@@ -1,81 +1,0 @@
-import { Component, HostListener } from '@angular/core'
-import {
-  PermissionsService,
-  PermissionAction,
-  PermissionType,
-} from 'src/app/services/permissions.service'
-import { SettingsService } from 'src/app/services/settings.service'
-import { ToastService } from 'src/app/services/toast.service'
-import { UploadDocumentsService } from 'src/app/services/upload-documents.service'
-
-@Component({
-  selector: 'pngx-file-drop',
-  templateUrl: './file-drop.component.html',
-  styleUrls: ['./file-drop.component.scss'],
-})
-export class FileDropComponent {
-  private fileLeaveTimeoutID: any
-  fileIsOver: boolean = false
-  hidden: boolean = true
-
-  constructor(
-    private settings: SettingsService,
-    private toastService: ToastService,
-    private uploadDocumentsService: UploadDocumentsService,
-    private permissionsService: PermissionsService
-  ) {}
-
-  public get dragDropEnabled(): boolean {
-    return (
-      this.settings.globalDropzoneEnabled &&
-      this.permissionsService.currentUserCan(
-        PermissionAction.Add,
-        PermissionType.Document
-      )
-    )
-  }
-
-  @HostListener('dragover', ['$event ']) onDragOver(event: DragEvent) {
-    if (!this.dragDropEnabled) return
-    event.preventDefault()
-    event.stopImmediatePropagation()
-    this.settings.globalDropzoneActive = true
-    // allows transition
-    setTimeout(() => {
-      this.fileIsOver = true
-    }, 1)
-    this.hidden = false
-    // stop fileLeave timeout
-    clearTimeout(this.fileLeaveTimeoutID)
-  }
-
-  @HostListener('dragleave', ['$event']) public onDragLeave(
-    event: DragEvent,
-    immediate: boolean = false
-  ) {
-    if (!this.dragDropEnabled) return
-    event.preventDefault()
-    event.stopImmediatePropagation()
-    this.settings.globalDropzoneActive = false
-
-    const ms = immediate ? 0 : 500
-
-    this.fileLeaveTimeoutID = setTimeout(() => {
-      this.fileIsOver = false
-      // await transition completed
-      setTimeout(() => {
-        this.hidden = true
-      }, 150)
-    }, ms)
-  }
-
-  @HostListener('drop', ['$event']) public onDrop(event: DragEvent) {
-    if (!this.dragDropEnabled) return
-    event.preventDefault()
-    event.stopImmediatePropagation()
-    this.onDragLeave(event, true)
-    this.uploadDocumentsService.uploadFiles(event.dataTransfer.files)
-    if (event.dataTransfer.files.length)
-      this.toastService.showInfo($localize`Initiating upload...`, 3000)
-  }
-}
