+++ src-ui/src/app/services/upload-documents.service.ts
@@ -1,5 +1,6 @@
 import { Injectable } from '@angular/core'
 import { HttpEventType } from '@angular/common/http'
+import { FileSystemFileEntry, NgxFileDropEntry } from 'ngx-file-drop'
 import {
   ConsumerStatusService,
   FileStatusPhase,
@@ -18,50 +19,61 @@ export class UploadDocumentsService {
     private consumerStatusService: ConsumerStatusService
   ) {}
 
+  onNgxFileDrop(files: NgxFileDropEntry[]) {
+    for (const droppedFile of files) {
+      if (droppedFile.fileEntry.isFile) {
+        const fileEntry = droppedFile.fileEntry as FileSystemFileEntry
+        fileEntry.file((file: File) => this.uploadFile(file))
+      }
+    }
+  }
+
   uploadFiles(files: FileList) {
     for (let index = 0; index < files.length; index++) {
-      const file = files.item(index)
-
-      let formData = new FormData()
-      formData.append('document', file, file.name)
-      let status = this.consumerStatusService.newFileUpload(file.name)
-
-      status.message = $localize`Connecting...`
-
-      this.uploadSubscriptions[file.name] = this.documentService
-        .uploadDocument(formData)
-        .subscribe({
-          next: (event) => {
-            if (event.type == HttpEventType.UploadProgress) {
-              status.updateProgress(
-                FileStatusPhase.UPLOADING,
-                event.loaded,
-                event.total
-              )
-              status.message = $localize`Uploading...`
-            } else if (event.type == HttpEventType.Response) {
-              status.taskId = event.body['task_id']
-              status.message = $localize`Upload complete, waiting...`
-              this.uploadSubscriptions[file.name]?.complete()
-            }
-          },
-          error: (error) => {
-            switch (error.status) {
-              case 400: {
-                this.consumerStatusService.fail(status, error.error.document)
-                break
-              }
-              default: {
-                this.consumerStatusService.fail(
-                  status,
-                  $localize`HTTP error: ${error.status} ${error.statusText}`
-                )
-                break
-              }
-            }
-            this.uploadSubscriptions[file.name]?.complete()
-          },
-        })
+      this.uploadFile(files.item(index))
     }
+  }
+
+  private uploadFile(file: File) {
+    let formData = new FormData()
+    formData.append('document', file, file.name)
+    let status = this.consumerStatusService.newFileUpload(file.name)
+
+    status.message = $localize`Connecting...`
+
+    this.uploadSubscriptions[file.name] = this.documentService
+      .uploadDocument(formData)
+      .subscribe({
+        next: (event) => {
+          if (event.type == HttpEventType.UploadProgress) {
+            status.updateProgress(
+              FileStatusPhase.UPLOADING,
+              event.loaded,
+              event.total
+            )
+            status.message = $localize`Uploading...`
+          } else if (event.type == HttpEventType.Response) {
+            status.taskId = event.body['task_id']
+            status.message = $localize`Upload complete, waiting...`
+            this.uploadSubscriptions[file.name]?.complete()
+          }
+        },
+        error: (error) => {
+          switch (error.status) {
+            case 400: {
+              this.consumerStatusService.fail(status, error.error.document)
+              break
+            }
+            default: {
+              this.consumerStatusService.fail(
+                status,
+                $localize`HTTP error: ${error.status} ${error.statusText}`
+              )
+              break
+            }
+          }
+          this.uploadSubscriptions[file.name]?.complete()
+        },
+      })
   }
 }
